feat: improve edit visuals
