+++ src/components/notecard.module.scss
@@ -2,7 +2,7 @@
   position: relative;
   color: white;
   height: auto;
-  margin: 0.5rem;
+  margin-bottom: 1rem;
   padding: 0.5rem;
   border-radius: 0.5rem;
   transition: padding 0.2s;
