+++ src/components/page.module.scss
@@ -4,106 +4,4 @@
   font-size: 1.5rem;
   height: 100%;
   overflow: hidden;
-
-  &.editHidden {
-    .textContainer {
-      height: 100%;
-    }
-
-    .textEdit {
-      height: 0;
-    }
-  }
-
-  .textContainer {
-    padding: 1rem 0;
-    position: relative;
-    line-height: 2.5ch;
-    height: 50%;
-    display: flex;
-    flex-direction: column;
-
-    .titleContainer {
-      height: 4rem;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      flex-shrink: 0;
-      padding: 0 0.5rem;
-      gap: 0.5rem;
-
-      background-color: rgb(139, 207, 228);
-      border-radius: 1000px;
-
-      span {
-        cursor: pointer;
-      }
-
-      h1 {
-        height: 100%;
-        overflow-y: hidden;
-        overflow-x: auto;
-        white-space: nowrap;
-        display: flex;
-        align-items: center;
-        margin: 0;
-      }
-    }
-
-    .text {
-      height: 100%;
-      overflow-y: auto;
-      overflow-x: hidden;
-    }
-  }
-
-  .textEdit {
-    position: relative;
-    padding: 1rem 0;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 50%;
-
-    label {
-      height: fit-content;
-      display: grid;
-      font-size: 1.5rem;
-      grid-template-columns: 1fr 3fr;
-      text-align: right;
-      gap: 1rem;
-
-      input {
-        font-size: 1rem;
-        padding: 0.5rem;
-        min-width: 0;
-      }
-    }
-
-    textarea {
-      padding: 0.5rem;
-      height: 70%;
-      width: 100%;
-      font-size: 1rem;
-      overflow-y: auto;
-    }
-  }
-
-  .reference {
-    color: white;
-    padding: 0 0.5rem;
-    border-radius: 0.5rem;
-    white-space: nowrap;
-    cursor: pointer;
-  }
-}
-
-@media (max-width: 800px) {
-  .notesContainer .textContainer {
-    height: 4rem;
-  }
-
-  .textEdit {
-    height: 100%;
-  }
 }
