+++ src/components/pageedit.module.scss
@@ -1,0 +1,54 @@
+.pageEdit {
+  position: relative;
+  padding: 1rem 0;
+  display: flex;
+  flex-direction: column;
+  gap: .5rem;
+  width: 100%;
+  height: 50%;
+
+  border-top: 1px solid black;
+
+  &.height0 {
+    height: 0;
+    padding: 0;
+    overflow: hidden;
+    border-top: 0;
+  }
+
+  label {
+    height: fit-content;
+    display: grid;
+    font-size: 1rem;
+    grid-template-columns: 1fr 3fr;
+    align-items: center;
+    text-align: right;
+    gap: 1rem;
+
+    input {
+      font-size: 1rem;
+      padding: 0.5rem;
+      min-width: 0;
+      border: 0;
+      background-color: white;
+      border-radius: 1rem;
+    }
+  }
+
+  textarea {
+    padding: 0.5rem;
+    height: 70%;
+    width: 100%;
+    font-size: 1rem;
+    overflow-y: auto;
+    border: 0;
+    background-color: white;
+    border-radius: 1rem;
+  }
+}
+
+@media (max-width: 800px) {
+  .pageEdit {
+    height: 100%;
+  }
+}
